Refactor PostCreator folder.
