+++ src/screens/Feed/components/PostCreator/components/GIFImage/index.js
@@ -1,0 +1,1 @@
+export * from './component';
