+++ src/screens/Feed/components/PostCreator/components/MediaItem/index.js
@@ -1,0 +1,1 @@
+export * from './component';
